% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{add.edges}
\alias{add.edges}
\title{Add edges}
\usage{
add.edges(per, E_List, S_List)
}
\description{
Add edges for iterative refitting.
}
\details{
Intended as internal function.
}
\author{
Andrew Kramer

Alexander Mercier
}
\keyword{internal}
