% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomNetworkGen.R
\name{ConfigNetwork_gen}
\alias{ConfigNetwork_gen}
\title{Configuration Model Generation}
\usage{
ConfigNetwork_gen(k, weights, w.prob = NULL)
}
\arguments{
\item{k}{Degree list \cr
##The degrees must sum to an even number}

\item{weights}{A vector of edge weights for configuration model to sample from}

\item{w.prob}{A vector of edge weight probabilities to sample edge weights with}
}
\value{
An adjacency matrix, \code{A}, of configuration network
}
\description{
Generate a configuration model from a given degree sequence
}
