% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{DFS}
\alias{DFS}
\title{Depth First search}
\usage{
DFS(Adj, v, discovered)
}
\arguments{
\item{Adj}{Adjacency matrix.}

\item{v}{Node to perform DFS from.}

\item{discovered}{A list of discovered nodes from \code{v}. If initializing the search, should be all FALSE.}
}
\value{
Logical of length n where TRUE denotes connected to node \code{v}.
}
\description{
Iterative depth first search.
}
\details{
Intended as internal function.
}
\author{
Andrew Kramer

Alexander Mercier
}
\keyword{internal}
