% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpielmanSrivastava_Sparse.R
\name{EffR2}
\alias{EffR2}
\title{Effective Resistance Approximation with RAM Features}
\usage{
EffR2(network, epsilon, n = NULL)
}
\arguments{
\item{network}{An adjacency matrix or edge list of the network; edge list should be in the format |in|out|weight|}

\item{epsilon}{Parameter to control resolution of the approximation}

\item{n}{The number of nodes; default is \code{NULL} and should not be changed unless the network is disconnected}
}
\value{
A matrix, \code{R}, such that \code{R_ij} gives the approximate effective resistance between node \code{i} and \code{j}
}
\description{
Approximately calculate the effective resistances between all nodes of a graph. \cr
Based on algorithm from Spielman-Srivastava (2011) \cr
This version of the effective resistance calculator is more RAM efficient, but slower than \link[simplifyNet]{EffR}; if speed is a concern and parallel processing is possible, \link[simplfyNet]{EffRPar} is likely a better alternative
}
