% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpielmanSrivastava_Sparse.R
\name{EffR_List}
\alias{EffR_List}
\title{Effective Resistance Matrix to Effective Resistance Edge List}
\usage{
EffR_List(R, network, n = NULL)
}
\arguments{
\item{R}{Matrix of effective resistance values}

\item{network}{Adjacency matrix of the corresponding network}

\item{n}{The number of nodes; default is \code{NULL} and should not be changed unless the network is disconnected}
}
\value{
Vector, \code{R_v}, of effective resistance values
}
\description{
Create a vector of effective resistance edge list from the output of \link[simplifyNet]{EffR}, or other versions
}
