% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Mod.R
\name{EList_Mtrx}
\alias{EList_Mtrx}
\title{Edge list to adjacency matrix}
\usage{
EList_Mtrx(E_List, n)
}
\arguments{
\item{E_List}{Edge list formatted |in|out|weight|}

\item{n}{Number of nodes}
}
\value{
Adjacency matrix constructed from edge list \code{E_List}
}
\description{
Convert an edge list to an adjacency matrix
}
