% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{find.neighbors}
\alias{find.neighbors}
\title{Find Neighbors}
\usage{
find.neighbors(Adj, v)
}
\arguments{
\item{Adj}{Adjacency matrix}

\item{v}{Node to find the neighbors of.}
}
\value{
Integers designating node indices of the adjacency matrix for the neighbors of \code{v}.
}
\description{
Finds the neighbors of a given node, \code{v}.
}
\details{
Intended as internal function.
}
\author{
Andrew Kramer

Alexander Mercier
}
\keyword{internal}
