% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{is.connected}
\alias{is.connected}
\title{Connectivity of Graph}
\usage{
is.connected(Adj)
}
\arguments{
\item{Adj}{Adjacency matrix.}
}
\value{
Return TRUE if network is connected and FALSE if not connected.If the network is directed, then this function checks if the network is strongly connected.
}
\description{
Tests the connectivity of a graph by performing a Depth First Search (DFS) from a random node.
}
\details{
Intended as internal function.
}
\author{
Andrew Kramer

Alexander Mercier
}
\keyword{internal}
