% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffRSparse.R
\name{Lap}
\alias{Lap}
\title{Graph laplacian calculator}
\usage{
Lap(A)
}
\arguments{
\item{A}{Weighted adjacency matrix of the class dgCMatrix}
}
\value{
Graph laplacian, \code{L}, of the class dgCMatrix.
}
\description{
Find the graph laplacian from a weighted adjacency matrix
}
