% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{LeverageScore}
\alias{LeverageScore}
\title{Total statistical leverage score}
\usage{
LeverageScore(E_List)
}
\arguments{
\item{E_List}{Edge list of the given network in the format of | node 1 | node 2 | weight |}
}
\value{
Returns the sum of all statistical leverage scores of each edge, \eqn{\sum_{e \in E} w_e \times R_e}
}
\description{
Computes the global statistical leverage score, \eqn{w_e\times R_e}, of a given network
}
