% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Mod.R
\name{Mtrx_EList}
\alias{Mtrx_EList}
\title{Adjacency matrix to edge list}
\usage{
Mtrx_EList(A)
}
\arguments{
\item{A}{Adjacency matrix}
}
\value{
An edge list \code{E_List} of adjacency matrix \code{A}.
}
\description{
Convert an adjacency matrix to an edge list
}
\author{
Author: Alexander M. Mercier
}
