% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{irefit}
\alias{irefit}
\title{Iterative refitting}
\usage{
irefit(E_List, func, tol, rank = "weight", connected = TRUE, per = 0.5)
}
\arguments{
\item{E_List}{Edge list of the given network in the format of | node 1 | node 2 | weight |.}

\item{func}{Model function whose input is the network and whose output is a single real value or a list of reevaluated weights in the first index and a real value in the second.\cr
A wrapper function may have to be written.}

\item{tol}{Allowed error around the original output of \code{func} approximated by the sparsified network within which edges are removed. Specifies if method converges.}

\item{rank}{Ranking of edges. Lower ranked edges are removed first. Must be the same length as \code{nrow(E_List)}.}

\item{connected}{If TRUE, connectivity of the network is prioritized over scoring by \code{func}.}

\item{per}{Percentage of edges to add/remove from the sparsifier at each step.}
}
\value{
Sparsified network, \code{func(H)}, which still maintains evaluator function, \code{func}, plus/minus \code{tol}.
}
\description{
Iterative sparsifcation based
}
\author{
Alexander Mercier
}
