% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffRSparse.R
\name{Mtrx_EList}
\alias{Mtrx_EList}
\title{Adjacency matrix to edge list}
\usage{
Mtrx_EList(A, directed = FALSE)
}
\arguments{
\item{A}{Weighted adjacency matrix.}

\item{directed}{Specifies if the network is directed or undirected. Default is set to undirected.}
}
\value{
An edge list, \code{E_List}, of adjacency matrix, \code{A}, of the form | n1 | n2 | weight |.
}
\description{
Convert an adjacency matrix to an edge list.
}
\author{
Alexander Mercier
}
