% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpielmanSrivastava_Sparse.R
\name{normprobs}
\alias{normprobs}
\title{Normalize Probabilities}
\usage{
normprobs(P)
}
\arguments{
\item{P}{List of probabilities}
}
\value{
A list of probabilities, \code{P_n}, such that they sum to 1
}
\description{
Normalize a list of probabilities such that \code{sum(probs)=1}
}
