% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomNetworkGen.R
\name{ER_gen}
\alias{ER_gen}
\title{Erdos-Renyi Random Network Generation}
\usage{
ER_gen(n, p, weights, w.prob = NULL, self.edge = FALSE)
}
\arguments{
\item{n}{The number of nodes in the Erdos-Renyi model}

\item{p}{Probability of linking two nodes with an edge}

\item{weights}{A vector of edge weights for Erdos-Renyi model to sample from}

\item{w.prob}{Vector of probabilities to sample edge weights with}

\item{self.edge}{To include self edges within the Erdos-Renyi model}
}
\value{
An adjacency matrix, \code{A}, of the Erdos-Renyi model
}
\description{
Generate an Erdos-Renyi model with specified edge weights and edge weight probabilities
}
\author{
Author: Alexander M. Mercier
}
