% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomNetworkGen.R
\name{RanGeo_gen}
\alias{RanGeo_gen}
\title{Random Geometric Network Generation}
\usage{
RanGeo_gen(n, r, weights, w.prob = NULL)
}
\arguments{
\item{n}{The number of nodes in the random geometric network}

\item{r}{Maximum euclidean distance for adding an edge in the metric space \code{[0,1)}}

\item{weights}{Edge weights for random geometric network to sample from}

\item{w.prob}{Vector of probabilities to sample edge weights with}
}
\value{
The adjacency matrix, \code{A}, of the random matrix with a data.frame, \code{pos}, of the x,y coordinates for each node
}
\description{
Generate a random geometric network with specified maximum edge distance
}
