% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomNetworkGen.R
\name{SBM_gen}
\alias{SBM_gen}
\title{Stochastic Block Model Generation}
\usage{
SBM_gen(s, W, weights, w.prob = NULL, self.edge = FALSE)
}
\arguments{
\item{s}{List of community sizes}

\item{W}{Matrix of probabilities such that \code{W_ij} is the probability of connecting community \code{i} and \code{j} with an edge}

\item{weights}{Vector of edge weights for the stochastic block model to sample from}

\item{w.prob}{Vector of probabilities to sample edge weights with}

\item{self.edge}{To include self edges within the Erdos-Renyi model}
}
\value{
An adjacency matrix, \code{A}, of a stochastic block model
}
\description{
Generate a stochastic block model from specified community sizes, between community probability of an edge, and within community probability of an edge
}
