% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{sparse.step}
\alias{sparse.step}
\title{Iterative refitting sparsification step}
\usage{
sparse.step(
  E_List,
  S_List,
  stepsize,
  spr_score,
  org_score,
  func,
  tol,
  per,
  connected
)
}
\value{
Return sparsified edge list.
}
\description{
Recursive sparsification step for iterative refitting.
}
\details{
Intended as internal function.
}
\author{
Andrew Kramer

Alexander Mercier
}
\keyword{internal}
