% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpielmanSrivastava_Sparse.R
\name{Qgen}
\alias{Qgen}
\alias{QGen}
\title{Random Bernoulli Matrix Generation}
\usage{
QGen(m, epsilon)
}
\arguments{
\item{m}{Number edges in the given network}

\item{epsilon}{Used to specify resolution of the approximation with regard to the Johnson-Lindenstrauss lemma}
}
\value{
Generates a random +/- 1/sqrt(k) Bernoulli matrix, \code{Q}
}
\description{
Compute random +/-\code{1/sqrt(k)} Bernoulli matrix where \code{k=24log(n)/epsilon^2} \cr
Used to fulfill Johnson-Lindenstrauss lemma \cr
From Spielman and Srivastava 2011 \cr
}
\author{
Author: Alexander M. Mercier
}
