% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffRSparse.R
\name{sqrt_WDiag}
\alias{sqrt_WDiag}
\title{Diagonal, m by m, weight matrix with \code{sqrt(w_e)} on the diagonal.}
\usage{
sqrt_WDiag(weights)
}
\arguments{
\item{weights}{List of edges corresponding to edge list.}
}
\value{
Return diagonal weight matrix, \code{W}, with \eqn{\sqrt(w_e)} on the diagonal.
}
\description{
Calculate the m by m diagonal weight matrix, \code{W}, for an inputted graph with \code{sqrt(w_e)} on the diagonal.
}
\details{
Intended as internal function.
}
\author{
Alexander Mercier
}
\keyword{internal}
