% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToivonenBestPath.R
\name{toivonen}
\alias{toivonen}
\title{Sparsification via Best Path}
\usage{
toivonen(E_List, directed = FALSE, associative = TRUE)
}
\arguments{
\item{E_List}{Edge list of the given network in the format of | node 1 | node 2 | weight |.}

\item{directed}{Specifies if the network is directed or undirected. Default is set to undirected.}

\item{associative}{Designates if the network is associative where edge weight determines "similarity" or "strength" or dissociative where edge weight denotes "dissimilarity" or "distance".\cr
If the network is associative, then the shortest path would be found by looking at \code{w_e^-1} where weaker association between nodes suggests a larger distance between nodes for shortest paths.
If the network is dissociative, then the shortest path would be between \code{w_e}.}
}
\value{
Edge list of sparsified network
}
\description{
Calcualtes network sparsifier from best path
}
\author{
Author: Alexander M. Mercier
}
