% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Mod.R
\name{WDiag}
\alias{WDiag}
\title{Weight Diagonal Matrix}
\usage{
WDiag(network)
}
\arguments{
\item{network}{Adjacency matrix or edge list of a network; edge list should be formatted |in|out|weight|}
}
\value{
Matrix, \code{W}, with network weights on the diagonal
}
\description{
Compute the weight diagonal matrix of a network
}
\examples{
network <- ER_gen(n=100, p=0.1, 1)
W = WDiag(network)
}
