% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpielmanSrivastava_Sparse.R
\name{EffR}
\alias{EffR}
\title{Effective Resistance Approximation}
\usage{
EffR(network, epsilon, n = NULL)
}
\arguments{
\item{network}{An adjacency matrix or edge list of the network; edge list should be in the format |in|out|weight|}

\item{epsilon}{Parameter to control resolution of the approximation}

\item{n}{The number of nodes; default is \code{NULL} and should not be changed unless the network is disconnected}
}
\value{
A matrix, \code{R}, such that \code{R_ij} gives the approximate effective resistance between node \code{i} and \code{j}
}
\description{
Approximately calculate the effective resistances between all nodes of a graph. \cr
Based on algorithm from Spielman-Srivastava (2011) \cr
This version of the effective resistance approximation calculator is RAM intensive. If you are working with a large network, use either \link[simplifyNet]{EffR2} or \link[simplifyNet]{EffRPar}
}
\examples{
A <- ER_gen(n=100, p=0.1, weights=1)
R <- EffR(network=A, epsilon=0.1)
}
