% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffRSparse.R
\name{EffRSparse}
\alias{EffRSparse}
\title{Sparsification through sampling effective resistances}
\usage{
EffRSparse(n, E_List, q, effR)
}
\arguments{
\item{n}{The number of nodes in the network.}

\item{E_List}{Edge list formatted | n1 | n2 | weight |.}

\item{q}{The numbers of samples taken. The fidelity to the original network increases as the number of samples increases, but decreases the sparseness.}

\item{effR}{Effective resistances corresponding to each edge. Should be in the same order as "weight".}
}
\value{
A sparsified network, \code{H}, edge list where the number of edges is dependent on the number of samples taken, \code{q}.
}
\description{
Sparsify an undirected network by sampling edges proportional to w_e * R_e.\cr
Approximately preserves the graph Laplacian, \code{L}, with increasing fidelity as the number of samples taken increases.
}
