% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpielmanSrivastava_Sparse.R
\name{EffRSparse}
\alias{EffRSparse}
\title{Effective Resistance Spectral Sparsification}
\usage{
EffRSparse(network, q, R, n = NULL)
}
\arguments{
\item{network}{Adjacency matrix of the given network}

\item{q}{Number of samples to independently sample with replacement}

\item{R}{Matrix of effective resistances; see \link[simplifyNet]{EffR}}

\item{n}{The number of nodes; default is \code{NULL} and should not be changed unless the network is disconnected}
}
\value{
Generate an effective resistance spectral sparsifer adjacency matrix, \code{H}
}
\description{
Create a effective resistance sparsifer \cr
From Spielman and Srivastava 2008 \cr
Effective resistance edge values are used to randomly sample edges with replacement. After \code{O(nlog^c(n)/epsilon^2)} samples with replacement, the algorithm produces a weighted
subgraph such that the Laplacian is approximately multiplicity preserved within +/-\code{epsilon}.
}
\examples{
A <- ER_gen(n=100, p=0.1, weights=1)
R <- EffRPar(network=A, epsilon=0.1, workers=16)
m = length(Mtrx_EList(A))
q = ceiling((m*(log(m))^1)/(0.1^2))
G <- EffRSparse(network=A, q=q, R=R)
}
