% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffRSparse.R
\name{igraph.to.elist}
\alias{igraph.to.elist}
\title{igraph to EList}
\usage{
igraph.to.elist(g)
}
\value{
Return an edge list with weights from an igraph object.
}
\description{
Convert igraph object to edge list.
}
\details{
Intended as internal function.
}
\author{
Alexander Mercier
}
\keyword{internal}
