% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffRSparse.R
\name{Lap}
\alias{Lap}
\title{Graph Laplacian calculator}
\usage{
Lap(A)
}
\arguments{
\item{A}{Weighted adjacency matrix.}
}
\value{
Graph Laplacian, \code{L}, of the class dgCMatrix.
}
\description{
Find the graph Laplacian from a weighted adjacency matrix.
}
\details{
Intended as internal function.
}
\examples{
A = matrix(c(0,1,1,1,0,1,1,1,0), 3 ,3)
L = Lap(A)
}
\author{
Alexander Mercier
}
\keyword{internal}
