% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Mod.R
\name{Mtrx_EList}
\alias{Mtrx_EList}
\title{Adjacency matrix to edge list}
\usage{
Mtrx_EList(A)
}
\arguments{
\item{A}{Adjacency matrix}
}
\value{
An edge list \code{E_List} of adjacency matrix \code{A}.
}
\description{
Convert an adjacency matrix to an edge list
}
\examples{
A <- ER_gen(n=100, p=0.1, 1)
E = Mtrx_EList(A)
}
\author{
Author: Alexander M. Mercier
}
