% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffRSparse.R
\name{normProbs}
\alias{normProbs}
\title{Normalize probabilities}
\usage{
normProbs(P)
}
\arguments{
\item{P}{Numerics to be normalized to probabilities.}
}
\value{
Return probabilities where their sum is 1.
}
\description{
Normalize numerics to probabilities such that the sum of the vector equals 1.
}
\details{
Intended as internal function.
}
\author{
Alexander Mercier
}
\keyword{internal}
