% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomNetworkGen.R
\name{RanGeo_gen}
\alias{RanGeo_gen}
\title{Random Geometric Network Generation}
\usage{
RanGeo_gen(n, r, weights, w.prob = NULL)
}
\arguments{
\item{n}{The number of nodes in the random geometric network}

\item{r}{Maximum euclidean distance for adding an edge in the metric space \code{[0,1)}}

\item{weights}{Edge weights for random geometric network to sample from}

\item{w.prob}{Vector of probabilities to sample edge weights with}
}
\value{
The adjacency matrix, \code{A}, of the random matrix with a data.frame, \code{pos}, of the x,y coordinates for each node
}
\description{
Generate a random geometric network with specified maximum edge distance
}
\examples{
n <- 100
r <- 0.15
weights <- c(0.25, 0.5, 0.75, 1)
w.prob <- c(0.50, 0.25, 0.15, 0.10)
RanGeo = RanGeo_gen(n=n, r=r, weights=weights, w.prob=w.prob)
A = RanGeo[[1]]
pos = RanGeo[[2]]
}
