% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{remove.edges}
\alias{remove.edges}
\title{Remove edges}
\usage{
remove.edges(per, E_List, S_List)
}
\value{
Return a truncated edge list with a decreased number of edges determined by \code{per}.
}
\description{
remove edges for iterative refitting.
}
\details{
Intended as internal function.
}
\author{
Andrew Kramer

Alexander Mercier
}
\keyword{internal}
