% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{rerank}
\alias{rerank}
\title{Rerank edges}
\usage{
rerank(E_List, rank)
}
\value{
Return an edge list reordered according to \code{rank}.
}
\description{
Rerank edges for iterative refitting.
}
\details{
Intended as internal function.
}
\author{
Andrew Kramer

Alexander Mercier
}
\keyword{internal}
