% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Mod.R
\name{sVIM}
\alias{sVIM}
\title{Vertex Incidence Matrix}
\usage{
sVIM(network)
}
\arguments{
\item{network}{Adjacency matrix or edge list with weights}
}
\value{
A vertex incidence matrix, \code{B},  made up of vertex incidence vectors
}
\description{
Compute vertex incidence matrix of a given network
}
\examples{
network <- ER_gen(n=100, p=0.1, 1)
B = sVIM(network)
}
