% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffRSparse.R
\name{sVIM}
\alias{sVIM}
\title{Signed vertex incidence matrix}
\usage{
sVIM(E_List)
}
\arguments{
\item{E_List}{Edge list formatted | n1 | n2 | weight |.}
}
\value{
Return a signed, m by n, vertex incidence matrix, \code{B}.
}
\description{
Calculate the signed vertex incidence matrix.
}
\details{
Intended as internal function.
}
\examples{
E_List = matrix(c(1,1,2,2,3,3,1,1,1), 3, 3)
colnames(E_List) <- c("n1", "n2", "weight")
B = sVIM(E_List)
}
\author{
Alexander Mercier
}
\keyword{internal}
