% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffRSparse.R
\name{WDiag}
\alias{WDiag}
\title{Diagonal, m by m, weight matrix}
\usage{
WDiag(weights)
}
\arguments{
\item{weights}{List of edges corresponding to edge list.}
}
\value{
Return diagonal weight matrix, \code{W}.
}
\description{
Calculate the m by m diagonal weight matrix, \code{W}, for an imputed graph.
}
\details{
Intended as internal function.
}
\author{
Alexander Mercier
}
\keyword{internal}
